egt working

+++ src/Display/Nextion.cpp
@@ -94,7 +94,7 @@ void Nextion::updateDisplayData(AppData currentData) {
     sendCmd("odometer.txt=\"" + formatNumber(currentData.odometer) + "\"");
     sendCmd("tripA.txt=\"" + formatNumber(currentData.tripA) + "\"");
     sendCmd("tripB.txt=\"" + formatNumber(currentData.tripB) + "\"");
-    sendCmd("egt.val=" + (String)currentData.egt);
+    sendCmd("egt.val=" + (String)(int)currentData.egt);
 
     sendCmd("leftturn.aph=" + (String)(currentData.leftBlinker ? "127" : "0"));
     sendCmd("rightturn.aph=" + (String)(currentData.rightBlinker ? "127" : "0"));
